Add files via upload
This is what I've created so far. 
The hangman.java has the main method.
There is an issue with the guess characters that I must fix. You can also try to solve the full world with guess word but if you miss 
you have to re-start the whole game. If you get the word correct, you go to the next word. It continues to you guessed all the words in the arraylist.